+++ go/path_srv/internal/handlers/psdedupe.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 	"net"
 
-	"github.com/scionproto/scion/go/lib/common"
 	"github.com/scionproto/scion/go/lib/ctrl/path_mgmt"
 	"github.com/scionproto/scion/go/lib/infra"
 	"github.com/scionproto/scion/go/lib/infra/dedupe"
@@ -39,43 +38,14 @@ func (req *segReq) BroadcastKey() string {
 	return fmt.Sprintf("%s %s", req.segReq, req.server)
 }
 
-type psDeduper struct {
-	msger infra.Messenger
-}
-
-func (pd *psDeduper) segsRequestFunc(ctx context.Context,
-	request dedupe.Request) dedupe.Response {
-
-	req := request.(*segReq)
-	segs, err := pd.msger.GetSegs(ctx, req.segReq, req.server, req.id)
-	if err != nil {
-		return dedupe.Response{Error: err}
-	}
-	return dedupe.Response{Data: segs}
-}
-
-func NewDeduper(msger infra.Messenger) *dedupe.Deduper {
-	psd := &psDeduper{msger: msger}
-	return dedupe.New(psd.segsRequestFunc, 0, 0)
-}
-
-func (h *segReqHandler) getSegsFromNetwork(ctx context.Context,
-	req *path_mgmt.SegReq, server net.Addr, id uint64) (*path_mgmt.SegReply, error) {
-
-	responseC, cancelF := h.segsDeduper.Request(ctx, &segReq{
-		segReq: req,
-		server: server,
-		id:     id,
-	})
-	defer cancelF()
-	select {
-	case response := <-responseC:
-		if response.Error != nil {
-			return nil, response.Error
+func NewGetSegsDeduper(msger infra.Messenger) dedupe.Deduper {
+	requestFunc := func(ctx context.Context, request dedupe.Request) dedupe.Response {
+		req := request.(*segReq)
+		segs, err := msger.GetSegs(ctx, req.segReq, req.server, req.id)
+		if err != nil {
+			return dedupe.Response{Error: err}
 		}
-		return response.Data.(*path_mgmt.SegReply), nil
-	case <-ctx.Done():
-		return nil, common.NewBasicError("Context done while waiting for Segs",
-			ctx.Err())
+		return dedupe.Response{Data: segs}
 	}
+	return dedupe.New(requestFunc, 0, 0)
 }
